requirements.txt
python==3.8
pytorch==1.10.2
numpy==1.21.0
matplotlib==3.1.2